+++ src/sdk/PnP.Core/Services/Core/PnPContext.cs
@@ -318,6 +318,24 @@ namespace PnP.Core.Services
             await BatchClient.ExecuteBatch(batch).ConfigureAwait(false);
         }
 
+        /// <summary>
+        /// Sets an access token to be used for this context. Only works when the current configuration is using the <see cref="OAuthAccessTokenConfiguration"/>
+        /// </summary>
+        /// <param name="accessToken">Access token to set</param>
+        public void SetAccessToken(string accessToken)
+        {
+            if (string.IsNullOrEmpty(accessToken))
+            {
+                throw new ArgumentNullException(nameof(accessToken));
+            }
+
+            if (!(AuthenticationProvider.Configuration is OAuthAccessTokenConfiguration))
+            {
+                throw new ClientException(ErrorType.ConfigurationError, "You can only set the access token when you've configured this context to use the OAuthAccessTokenConfiguration");
+            }
+
+            (AuthenticationProvider.Configuration as OAuthAccessTokenConfiguration).AccessToken = accessToken;
+        }
         #endregion
 
 #if DEBUG